+++ Trebuchet/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### From Sprint #1901 (Radoub-level)
+
+- Fixed settings showing local log level instead of effective shared level (#1891)
+
 ---
 
 ## [1.24.0-alpha] - 2026-03-17